pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-build-bot"
version = "0.1.0"
description = "Telegram bot for structured construction progress reporting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22.0",
  "google-genai>=1.49.0,<2.0.0",
  "jsonschema>=4.22.0,<5.0.0",
  "gspread>=6.1.0,<7.0.0",
  "google-auth>=2.35.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "pytest-asyncio>=0.23.7,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
  "mypy>=1.10.0,<2.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
strict = true
